requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-import-export==3.3.7
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
et-xmlfile==1.1.0
filelock==3.13.4
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.6
MarkupPy==1.14
neverbounce-sdk==4.3.0
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==1.7.1
pytest==8.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.32.3
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
